temp workaround unnecessary fetching of CC
in CAPI 1.1.0+, clusterctl when processing the cluster template for a
specific plan (flavor), will also attempt to fetch CC from the cluster
or the provider repository (local filesystem in our case), but neither
are populated with the CC we use in ccgen dry-runs.

Working around this requires some hacks in CAPI, so this change repoints
go.mod to a temporary fork with workaround.

The workaround is activate via setting CLUSTERCTL_SKIP_FETCH_CC=true